Trick can now determine winning player.
Card has a helper function that supports the Trick in determining the winning
player by selecting the highest card from a given list of cards.

Fixed a bug in the isHigherThan() method of Card as it did not check all
possible states correctly. There are four states the method can be in, namely:
card is trump and other card is trump, card is trump and other card is not,
card is not trump and other card is, both cards are not trump. While
previously there were only two: both cards are trump, both cards are not.