Decreased the complexity of calculating the consecutive card score according to Joost's following comment:
Line-by-line comments:

File: /trunk/src/org/klaverjasaichallenge/score/RoemScore.java (r90)
===============================================================================

Line 76:                for(Card card : trick.getCards()) {
-------------------------------------------------------------------------------
Is the negative part necessary? Because you loop all the cards, you will also
encounter the card first in line, where differencePositive is true but
differenceNegative is false. When you remove the notion of differenceNegative
not all cards within a consecutive row will trigger; but only the first one.
This is however not a problem, and removing the notion of differenceNegative
will decrease the complexity.