Now using log4j instead of good old System.out.println.
This is useful since we can now distinguish between different levels of
logging. Currently two levels are used:
- DEBUG: To really show all output of every game that is played.
- INFO: To only show the end result of all games.
The INFO level is especially useful because when logging on this level, the
DEBUG messages are ignored and thus the program will run alot faster by not
printing the output of all (possibly) 10.000 games that will be played.